Ännu mer stora kliv framåt
script.js är ordentligt overhaulad. Vacker och utförligt kommenterad. Dock slutar spelet i förtid när man kör, så det måste jag lösa.

Är glad att jag hittade db.transaction

+++ db/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE users (
 CREATE TABLE scores (
     id INTEGER PRIMARY KEY,
     user_id INTEGER,
-    score INTEGER,
+    score REAL,
     date DATETIME,
     set_id INTEGER
 );
